refactor: move pkg/ to internal/ to enforce non-public API
logx is a CLI/kubectl plugin, not a library; nothing is meant for external
import. Moving pkg/* to internal/* lets the compiler enforce that boundary and
frees future refactoring.

- git mv pkg internal (history preserved)
- Update all 20 Go import paths
- Update ldflags version path in justfile and .goreleaser.yaml
  (validated: `just build` + `logx version --short` embeds the git version;
  `goreleaser check` passes)
