+++ internal/format/output.go
@@ -5,8 +5,8 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/dantech2000/logx/pkg/kubernetes"
-	"github.com/dantech2000/logx/pkg/terminal"
+	"github.com/dantech2000/logx/internal/kubernetes"
+	"github.com/dantech2000/logx/internal/terminal"
 	"gopkg.in/yaml.v2"
 )
