+++ internal/kubernetes/container.go
@@ -5,7 +5,7 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/dantech2000/logx/pkg/terminal"
+	"github.com/dantech2000/logx/internal/terminal"
 	"github.com/fatih/color"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
