+++ internal/kubernetes/testdata/expected/multiline-warn.txt
