+++ internal/kubernetes/testdata/expected/plaintext-warn.txt
