+++ internal/kubernetes/testdata/expected/traefik-warn.txt
