+++ internal/kubernetes/testdata/logs/json-mixed.txt
