+++ internal/kubernetes/testdata/logs/logfmt-mixed.txt
