+++ internal/kubernetes/testdata/logs/multiline-stacktrace.txt
