+++ internal/kubernetes/testdata/logs/plaintext-mixed.txt
