+++ internal/kubernetes/testdata/logs/traefik.txt
