+++ internal/kubernetes/testdata/real-imagepull/pod.yaml
