+++ internal/kubernetes/testdata/real-portal/logs.txt
