+++ internal/kubernetes/testdata/real-rails/events.yaml
