+++ internal/kubernetes/testdata/real-rails/logs.txt
