+++ internal/kubernetes/testdata/real-traefik/events.yaml
