+++ internal/kubernetes/testdata/real-traefik/pod.yaml
