+++ internal/logging/scanner.go
