+++ internal/terminal/sanitize.go
