+++ internal/terminal/sanitize_test.go
